improvement: restructured error and invite pages
