+++ apps/sim/app/invite/invite-error/page.tsx
@@ -1,0 +1,8 @@
+import InviteError from './invite-error'
+
+// Generate this page on-demand instead of at build time
+export const dynamic = 'force-dynamic'
+
+export default function InviteErrorPage() {
+  return <InviteError />
+}
